Rename files SVNW-99
