+++ src/shared/hooks/useThemeColor.ts
@@ -2,7 +2,7 @@
  * Learn more about light and dark modes:
  * https://docs.expo.dev/guides/color-schemes/
  */
-import { Colors } from '@/shared/constants/Colors';
+import { Colors } from '@/shared/constants/colors';
 import { useColorScheme } from 'react-native';
 
 export function useThemeColor(
